feat: Update index.html with favicon and font links
The code changes include adding the favicon link and font links to the index.html file. The favicon link specifies the icon to be displayed in the browser tab, while the font links import the "Josefin Sans" font from Google Fonts.

Based on the recent user commits and repository commits, it seems that the project is following a convention of using the "feat" type for feature-related changes. Therefore, a suitable commit message for these code changes would be "feat: Update index.html with favicon and font links".